spec: rename assert-helpers to expect-helpers
